+++ Avatar 2.0 VRCSDK/Dependencies/VRChat/Editor/ControlPanel/TBASdkChangelog.cs
@@ -59,8 +59,7 @@ UnityEditor.EditorPrefs.GetFloat("SDKColor_A")
         EditorGUILayout.EndHorizontal();
 
         EditorGUILayout.BeginVertical(boxGuiStyle);
-        EditorGUILayout.LabelField("The Black Arms Discord: https://discord.gg/m6UfHkY", EditorStyles.boldLabel);
-        EditorGUILayout.LabelField("DEDSEC Discord: https://discord.gg/r7RcJCv", EditorStyles.boldLabel);
+        EditorGUILayout.LabelField("The Black Arms Discord: theblackarms.cf", EditorStyles.boldLabel);
         EditorGUILayout.LabelField("The Black Arms SDX Dev: PhoenixAceVFX", EditorStyles.boldLabel);
         EditorGUILayout.LabelField("The Black Arms SDX Contributors", EditorStyles.boldLabel);
         EditorGUILayout.LabelField("Wolfie (Huge help with legal and github!)", EditorStyles.boldLabel);
@@ -70,6 +69,7 @@ UnityEditor.EditorPrefs.GetFloat("SDKColor_A")
         EditorGUILayout.LabelField("Plague (LoadBundle Script Creator)", EditorStyles.boldLabel);
         EditorGUILayout.LabelField("Controser (AntiYoink[In 3.0] and SENTINEL IMPORTER)", EditorStyles.boldLabel);
         EditorGUILayout.LabelField("Bunny (For their Utilities!)", EditorStyles.boldLabel);
+        EditorGUILayout.LabelField("Zombie2312 (Total VPS Hosting)", EditorStyles.boldLabel);
         EditorGUILayout.EndVertical();
 
         changesScroll = EditorGUILayout.BeginScrollView(changesScroll, GUILayout.Width(SdkWindowWidth));
@@ -85,6 +85,10 @@ Much easier to do than to make an entire remake
 This is completely free of charge as I refuse to make
 Any money or profit from my SDX and I will not rebrand
 My SDX, this decision is final and wont be changed.
+
+The Black Arms SDX 2.2
+Migrated EVERYTHING to a dedicated host!
+-Huge thanks to Zombie2312 for the hosting!
 
 The Black Arms SDX 2.1.1
 Added a few extra assets