+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 >Any money or profit from my SDX and I will not rebrand  
 >My SDX, this decision is final and wont be changed.  
 
+>The Black Arms SDX 2.2  
+>Migrated EVERYTHING to a dedicated host!  
+>-Huge thanks to Zombie2312 for the hosting!  
+
 >The Black Arms SDX 2.1.1  
 >Added some extra assets